Revert "Package division"
This reverts commit 069073b44ef30a56379556cec471a0eb2bf70f03.
