+++ MarvelHeroes/app/src/main/java/it/tiburtinavalley/marvelheroes/Activity/MainActivity.java
@@ -1,4 +1,4 @@
-package it.tiburtinavalley.marvelheroes.activity;
+package it.tiburtinavalley.marvelheroes.Activity;
 
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
@@ -24,11 +24,10 @@ import com.bumptech.glide.Glide;
 import java.util.ArrayList;
 import java.util.List;
 
-import it.tiburtinavalley.marvelheroes.model.HeroModel;
+import it.tiburtinavalley.marvelheroes.Model.HeroModel;
 import it.tiburtinavalley.marvelheroes.R;
-import it.tiburtinavalley.marvelheroes.recyclerviewadapters.HeroAdapter;
-import it.tiburtinavalley.marvelheroes.volley.ImageApiVolley;
-import it.tiburtinavalley.marvelheroes.volley.MarvelApiVolley;
+import it.tiburtinavalley.marvelheroes.Volley.ImageApiVolley;
+import it.tiburtinavalley.marvelheroes.Volley.MarvelApiVolley;
 
 public class MainActivity extends AppCompatActivity {
     MarvelApiVolley volleyMarvel;
@@ -60,7 +59,7 @@ public class MainActivity extends AppCompatActivity {
                 private void fillRecView(List<HeroModel> heroes) {
                     RecyclerView.LayoutManager layoutManager = new LinearLayoutManager(MainActivity.this);
                     rvHeroes.setLayoutManager(layoutManager);
-                    HeroAdapter mAdapter = new HeroAdapter(heroes, getApplicationContext());
+                    HeroAdapter mAdapter = new HeroAdapter(heroes);
                     rvHeroes.setAdapter(mAdapter);
                 }
             };
@@ -88,6 +87,67 @@ public class MainActivity extends AppCompatActivity {
         public void hideSoftKeyboard(Activity activity){
             InputMethodManager imm = (InputMethodManager)activity.getSystemService(Activity.INPUT_METHOD_SERVICE);
             imm.hideSoftInputFromWindow(activity.getCurrentFocus().getWindowToken(), 0);
+        }
+    }
+
+    private class HeroAdapter extends RecyclerView.Adapter<HeroAdapter.Holder> implements View.OnClickListener {
+        private final List<HeroModel> heroes;
+
+        HeroAdapter(List<HeroModel> all) {
+            heroes = new ArrayList<>();
+            heroes.addAll(all);
+        }
+
+        @NonNull
+        @Override
+        public Holder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
+            ConstraintLayout cl;
+            cl = (ConstraintLayout) LayoutInflater
+                    .from(parent.getContext())
+                    .inflate(R.layout.hero_layout, parent, false);
+            cl.setOnClickListener(this);
+            return new Holder(cl);
+        }
+
+        // This method sets the layout of the hero
+        @Override
+        public void onBindViewHolder(@NonNull Holder holder, int position) {
+            HeroModel hero = heroes.get(position);
+            holder.tvHeroName.setText(hero.getName());
+            imgVolley.addHeroImg(holder.ivHeroPic);
+            if (!hero.getThumbnail().getPath().equalsIgnoreCase("")
+                    && !hero.getThumbnail().getExtension().equalsIgnoreCase("")) {
+                String urlThumbnail = hero.getThumbnail().getPath().replaceFirst("http", "https")
+                        + "." + hero.getThumbnail().getExtension();
+                Glide.with(holder.itemView).load(urlThumbnail).into(holder.ivHeroPic);
+            }
+        }
+
+        @Override
+        public int getItemCount() {
+            return heroes.size();
+        }
+
+        //
+        @Override
+        public void onClick(View v) {
+            int position = ((RecyclerView) v.getParent()).getChildAdapterPosition(v);
+            HeroModel hero = heroes.get(position);
+
+            Intent i = new Intent(getApplicationContext(), HeroDetailActivity.class);
+            i.putExtra("hero", hero);
+            startActivity(i);
+        }
+
+        class Holder extends RecyclerView.ViewHolder {
+            final TextView tvHeroName;
+            final ImageView ivHeroPic;
+
+            Holder(@NonNull View itemView) {
+                super(itemView);
+                tvHeroName = itemView.findViewById(R.id.tvHeroName);
+                ivHeroPic = itemView.findViewById(R.id.ivHeroPic);
+            }
         }
     }
 }
