+++ MarvelHeroes/app/src/main/java/it/tiburtinavalley/marvelheroes/Model/Comics.java
@@ -1,8 +1,10 @@
-package it.tiburtinavalley.marvelheroes.model;
+package it.tiburtinavalley.marvelheroes.Model;
 
 import android.os.Parcel;
 import android.os.Parcelable;
 import java.util.List;
+
+import it.tiburtinavalley.marvelheroes.Volley.ImageApiVolley;
 
 /* Model per mantenere i dati relativi ai fumetti legati agli eroi */
