+++ MarvelHeroes/app/src/main/java/it/tiburtinavalley/marvelheroes/Model/Stories.java
@@ -1,4 +1,4 @@
-package it.tiburtinavalley.marvelheroes.model;
+package it.tiburtinavalley.marvelheroes.Model;
 
 import android.os.Parcel;
 import android.os.Parcelable;
