+++ MarvelHeroes/app/src/main/java/it/tiburtinavalley/marvelheroes/Volley/ComicsVolley.java
@@ -1,8 +1,7 @@
-package it.tiburtinavalley.marvelheroes.volley;
+package it.tiburtinavalley.marvelheroes.Volley;
 
 import android.content.Context;
 import android.util.Log;
-
 import com.android.volley.Request;
 import com.android.volley.RequestQueue;
 import com.android.volley.Response;
@@ -11,34 +10,37 @@ import com.android.volley.toolbox.StringRequest;
 import com.android.volley.toolbox.Volley;
 import com.google.gson.Gson;
 import com.google.gson.reflect.TypeToken;
-
 import org.json.JSONException;
 import org.json.JSONObject;
 import java.lang.reflect.Type;
 import java.util.List;
-import it.tiburtinavalley.marvelheroes.model.Stories;
+import it.tiburtinavalley.marvelheroes.Model.Comics;
+import it.tiburtinavalley.marvelheroes.Model.Series;
 
-public abstract class StoriesVolley implements Response.ErrorListener, Response.Listener<String>{
+/* in questa classe , vengono gestite le ricerche in Internet per cercare le informazioni
+   relative a Storie, Fumetti e Serie collegate agli eroi*/
+
+public abstract class ComicsVolley implements Response.ErrorListener, Response.Listener<String>{
     private String urlBase = "https://gateway.marvel.com/v1/public/characters/%s";
     private String apiKey = "ts=1&apikey=467ab31077a4aa2037776afb61241da4&hash=21f601a3255711a8d8bad803d062e9ea";
     private RequestQueue requestQueue;
 
-    public abstract void fillStories(List<Stories> storiesList);
+    public abstract void fillComics(List<Comics> comicsList);
 
-    public StoriesVolley(Context context){
+    public ComicsVolley(Context context){
         requestQueue = Volley.newRequestQueue(context);
     }
 
-    public void getStoriesInfo(String heroId){
-        String story = heroId + "/stories?";
-
+    public void getComicsInfo(String heroId){
+        String comic = heroId + "/comics?";
+        comicApiCall(comic);
     }
 
-    private void storiesApiCall(String storyUrl){
-        String url = urlBase + apiKey;
-        url = String.format(urlBase, storyUrl);
+    private void comicApiCall(String comicUrl){
+        String url = urlBase + apiKey; // usiamo una stringa di appoggio così da poter ripetere la chiamata
+        url = String.format(url, comicUrl);
         StringRequest sr = new StringRequest(Request.Method.GET,
-                urlBase,
+                url,
                 this,
                 this);
         requestQueue.add(sr);
@@ -56,12 +58,13 @@ public abstract class StoriesVolley implements Response.ErrorListener, Response.
         try {
             JSONObject jsonObject = new JSONObject(response);
             comics = jsonObject.getJSONObject("data").getJSONArray("results").toString();
-            Type listType = new TypeToken<List<Stories>>() {
+            Type listType = new TypeToken<List<Comics>>() {
             }.getType();
-            List<Stories> storiesList = gson.fromJson(comics, listType);
-            if (storiesList != null && storiesList.size() > 0) {
+            List<Comics> comicsList = gson.fromJson(comics, listType);
+            if (comicsList != null && comicsList.size() > 0) {
+                Log.w("CA", "" + comicsList.size());
                 //db.cocktailDAO().insertAll(cnt);    // NON OBBLIGATORIO
-                fillStories(storiesList);
+                fillComics(comicsList);
             }
         } catch (JSONException e) {
             e.printStackTrace();
