+++ MarvelHeroes/app/src/main/java/it/tiburtinavalley/marvelheroes/Volley/ImageApiVolley.java
@@ -1,4 +1,4 @@
-package it.tiburtinavalley.marvelheroes.volley;
+package it.tiburtinavalley.marvelheroes.Volley;
 
 import android.content.Context;
 import android.graphics.Bitmap;
