+++ MarvelHeroes/app/src/main/java/it/tiburtinavalley/marvelheroes/Volley/SeriesVolley.java
@@ -1,7 +1,8 @@
-package it.tiburtinavalley.marvelheroes.volley;
+package it.tiburtinavalley.marvelheroes.Volley;
 
 import android.content.Context;
 import android.util.Log;
+
 import com.android.volley.Request;
 import com.android.volley.RequestQueue;
 import com.android.volley.Response;
@@ -10,34 +11,35 @@ import com.android.volley.toolbox.StringRequest;
 import com.android.volley.toolbox.Volley;
 import com.google.gson.Gson;
 import com.google.gson.reflect.TypeToken;
+
 import org.json.JSONException;
 import org.json.JSONObject;
+
 import java.lang.reflect.Type;
 import java.util.List;
-import it.tiburtinavalley.marvelheroes.model.Comics;
 
-/* in questa classe , vengono gestite le ricerche in Internet per cercare le informazioni
-   relative a Storie, Fumetti e Serie collegate agli eroi*/
+import it.tiburtinavalley.marvelheroes.Model.Series;
+import it.tiburtinavalley.marvelheroes.Model.Stories;
 
-public abstract class ComicsVolley implements Response.ErrorListener, Response.Listener<String>{
+public abstract class SeriesVolley implements Response.ErrorListener, Response.Listener<String>{
     private String urlBase = "https://gateway.marvel.com/v1/public/characters/%s";
     private String apiKey = "ts=1&apikey=467ab31077a4aa2037776afb61241da4&hash=21f601a3255711a8d8bad803d062e9ea";
     private RequestQueue requestQueue;
 
-    public abstract void fillComics(List<Comics> comicsList);
+    public abstract void fillSeries(List<Series> seriesList);
 
-    public ComicsVolley(Context context){
+    public SeriesVolley(Context context){
         requestQueue = Volley.newRequestQueue(context);
     }
 
-    public void getComicsInfo(String heroId){
-        String comic = heroId + "/comics?";
-        comicApiCall(comic);
+    public void getStoriesInfo(String heroId){
+        String serie = heroId + "/series?";
+        seriesApiCall(serie);
     }
 
-    private void comicApiCall(String comicUrl){
-        String url = urlBase + apiKey; // usiamo una stringa di appoggio così da poter ripetere la chiamata
-        url = String.format(url, comicUrl);
+    private void seriesApiCall(String storyUrl){
+        String url = urlBase + apiKey;
+        url = String.format(url, storyUrl);
         StringRequest sr = new StringRequest(Request.Method.GET,
                 url,
                 this,
@@ -57,13 +59,12 @@ public abstract class ComicsVolley implements Response.ErrorListener, Response.L
         try {
             JSONObject jsonObject = new JSONObject(response);
             comics = jsonObject.getJSONObject("data").getJSONArray("results").toString();
-            Type listType = new TypeToken<List<Comics>>() {
+            Type listType = new TypeToken<List<Series>>() {
             }.getType();
-            List<Comics> comicsList = gson.fromJson(comics, listType);
-            if (comicsList != null && comicsList.size() > 0) {
-                Log.w("CA", "" + comicsList.size());
+            List<Series> seriesList = gson.fromJson(comics, listType);
+            if (seriesList != null && seriesList.size() > 0) {
                 //db.cocktailDAO().insertAll(cnt);    // NON OBBLIGATORIO
-                fillComics(comicsList);
+                fillSeries(seriesList);
             }
         } catch (JSONException e) {
             e.printStackTrace();
